astraos/backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# ── Google AI & ADK ──────────────────────────────────────────────────────
google-genai>=1.10.0            # Gemini API + Live streaming
google-adk>=1.0.0               # Agent Development Kit (LlmAgent, FunctionTool, Runner)

# ── Google Cloud ─────────────────────────────────────────────────────────
google-cloud-firestore>=2.19.0  # Firestore + vector search

# ── Google APIs (Gmail + Calendar) ───────────────────────────────────────
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0

# ── Utilities ─────────────────────────────────────────────────────────────
httpx>=0.27.0                   # Async HTTP client
aiofiles>=23.0.0                # Async file I/O
